# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SabiScore API",
    "description": "Football Prediction API with advanced analytics and odds comparison",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/matches/upcoming": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "Get Upcoming Matches",
        "description": "Get upcoming matches for predictions\n\nReturns matches scheduled within the next N days, optionally filtered by league.\nResults are cached for 5 minutes to optimize performance.",
        "operationId": "matches-get_upcoming_matches",
        "parameters": [
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by league (epl, bundesliga, laliga, etc)",
              "title": "League"
            },
            "description": "Filter by league (epl, bundesliga, laliga, etc)"
          },
          {
            "name": "days_ahead",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "description": "Number of days to look ahead",
              "default": 7,
              "title": "Days Ahead"
            },
            "description": "Number of days to look ahead"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of matches to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum number of matches to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/predictions/": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Create Prediction",
        "description": "Generate comprehensive match prediction with value bet analysis\n\nPipeline:\n1. Fetch historical data (180k+ matches)\n2. Engineer 220+ features\n3. Run ensemble models (RF, XGBoost, LightGBM)\n4. Apply Platt calibration\n5. Detect value bets (min edge: 4.2%)\n6. Calculate Smart Kelly stakes (\u215b Kelly)\n\nTarget: <150ms response time @ 10k CCU",
        "operationId": "predictions-create_prediction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeagueCode": {
        "type": "string",
        "enum": [
          "epl",
          "bundesliga",
          "la_liga",
          "serie_a",
          "ligue_1"
        ],
        "title": "LeagueCode",
        "description": "Supported leagues"
      },
...
      "MatchListResponse": {
        "properties": {
          "matches": {
            "items": {
              "$ref": "#/components/schemas/MatchSummary"
            },
            "type": "array",
            "title": "Matches"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "league_filter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League Filter"
          },
          "date_range_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range Days"
          }
        },
        "type": "object",
        "required": [
          "matches",
          "total"
        ],
        "title": "MatchListResponse",
        "description": "Container for a list of matches with simple summary items."
      },
      "MatchPredictionRequest": {
        "properties": {
          "match_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Id"
          },
          "home_team": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Home Team"
          },
          "away_team": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Away Team"
          },
          "league": {
            "$ref": "#/components/schemas/LeagueCode",
            "description": "League identifier"
          },
          "kickoff_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kickoff Time"
          },
          "odds": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Odds",
            "description": "Market odds (home_win, draw, away_win)"
          },
          "bankroll": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100000000.0,
                "minimum": 1000.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bankroll",
            "description": "Bankroll in Naira (\u20a6) for Kelly calculation",
            "default": 10000
          }
        },
        "type": "object",
        "required": [
          "home_team",
          "away_team",
          "league"
        ],
        "title": "MatchPredictionRequest",
        "description": "Request model for match prediction"
      },
...
      "MatchSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "league": {
            "type": "string",
            "title": "League"
          },
          "match_date": {
            "type": "string",
            "format": "date-time",
            "title": "Match Date"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "scheduled"
          },
          "has_odds": {
            "type": "boolean",
            "title": "Has Odds",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "home_team",
          "away_team",
          "league",
          "match_date"
        ],
        "title": "MatchSummary",
        "description": "Lean response tailored for listings and search results."
      },
...
      "PredictionResponse": {
        "properties": {
          "match_id": {
            "type": "string",
            "title": "Match Id"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "league": {
            "$ref": "#/components/schemas/LeagueCode"
          },
          "predictions": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Predictions",
            "description": "Probabilities for home_win, draw, away_win"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "Highest probability"
          },
          "brier_score": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Brier Score",
            "description": "Expected Brier score"
          },
          "value_bets": {
            "items": {
              "$ref": "#/components/schemas/ValueBetResponse"
            },
            "type": "array",
            "title": "Value Bets",
            "description": "Identified value betting opportunities"
          },
          "confidence_intervals": {
            "additionalProperties": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "object",
            "title": "Confidence Intervals",
            "description": "95% confidence intervals for each outcome"
          },
          "explanations": {
            "type": "object",
            "title": "Explanations",
            "description": "SHAP feature importance and explanations"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Model metadata and performance metrics"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "match_id",
          "home_team",
          "away_team",
          "league",
          "predictions",
          "confidence",
          "brier_score"
        ],
        "title": "PredictionResponse",
        "description": "Comprehensive match prediction response",
        "example": {
          "away_team": "Liverpool",
          "brier_score": 0.178,
          "confidence": 0.421,
          "created_at": "2025-11-11T14:32:00Z",
          "home_team": "Arsenal",
          "league": "epl",
          "match_id": "epl_2025_234",
          "metadata": {
            "league_model": "epl",
            "model_version": "3.0",
            "processing_time_ms": 142
          },
          "predictions": {
            "away_win": 0.293,
            "draw": 0.286,
            "home_win": 0.421
          },
          "value_bets": [
            {
              "edge_ngn": 186,
              "kelly_stake_ngn": 53720,
              "market": "home_win",
              "match_id": "epl_2025_234",
              "odds": 1.96
            }
          ]
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ValueBetResponse": {
        "properties": {
          "match_id": {
            "type": "string",
            "title": "Match Id",
            "description": "Match identifier"
          },
          "market": {
            "type": "string",
            "title": "Market",
            "description": "Market name, e.g. home_win"
          },
          "odds": {
            "type": "number",
            "minimum": 1.01,
            "title": "Odds",
            "description": "Offered decimal odds"
          },
          "fair_probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Fair Probability"
          },
          "implied_probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Implied Probability"
          },
          "edge_percent": {
            "type": "number",
            "title": "Edge Percent",
            "description": "Edge expressed as percentage"
          },
          "edge_ngn": {
            "type": "number",
            "title": "Edge Ngn",
            "description": "Edge translated to Naira per \u20a610k stake"
          },
          "kelly_stake_ngn": {
            "type": "number",
            "minimum": 0.0,
            "title": "Kelly Stake Ngn",
            "description": "Recommended stake in Naira"
          },
          "kelly_fraction": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Kelly Fraction",
            "description": "Kelly fraction applied"
          },
          "clv_ngn": {
            "type": "number",
            "title": "Clv Ngn",
            "description": "Estimated closing line value in Naira"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "expected_roi": {
            "type": "number",
            "title": "Expected Roi",
            "description": "Expected ROI for this value bet"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "pinnacle_close": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 1.01
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinnacle Close",
            "description": "Reference Pinnacle closing price, if available"
          }
        },
        "type": "object",
        "required": [
          "match_id",
          "market",
          "odds",
          "fair_probability",
          "implied_probability",
          "edge_percent",
          "edge_ngn",
          "kelly_stake_ngn",
          "kelly_fraction",
          "clv_ngn",
          "confidence",
          "expected_roi"
        ],
        "title": "ValueBetResponse",
        "description": "Rich description of a detected value betting opportunity.",
        "example": {
          "clv_ngn": 81.0,
          "confidence": 0.847,
          "created_at": "2025-11-11T14:32:00Z",
          "edge_ngn": 186.0,
          "edge_percent": 9.3,
          "expected_roi": 0.089,
          "fair_probability": 0.563,
          "implied_probability": 0.51,
          "kelly_fraction": 0.125,
          "kelly_stake_ngn": 53720.0,
          "market": "home_win",
          "match_id": "epl_2025_234",
          "odds": 1.96
        }
      },
...
    }
  },
...
}